Moved logger and Rabbit utils to other packages
