+++ Client/RabbitMq/__init__.py
@@ -1,0 +1,2 @@
+from . import Query
+from . import RabbitMQClient
