+++ Client/SizeFilterConsumer.py
@@ -1,10 +1,9 @@
 import json
 import logging
 
-from CustomLogFormatter import CustomLogFormatter
-from Query import ResultResponse
-from RabbitMQClient import RabbitMQSyncConsumer, RabbitMQProducer
-from SizeFilter import filter_by_KB
+from Logger.CustomLogFormatter import CustomLogFormatter
+from RabbitMq.Query import ResultResponse
+from RabbitMq.RabbitMQClient import RabbitMQSyncConsumer, RabbitMQProducer
 
 logger = logging.getLogger("SimpleFilterConsumer")
 logger.setLevel(logging.DEBUG)
@@ -21,6 +20,14 @@ def send_result(prod, result):
 
 
 if __name__ == '__main__':
+    import sys
+    from pathlib import Path  # if you haven't already done so
+
+    file = Path(__file__).resolve()
+    parent, root = file.parent, file.parents[1]
+    sys.path.append(str(root))
+    from Client.SizeFilter import filter_by_KB
+
     logger.info("Starting SizeFilterConsumer")
     consumer = RabbitMQSyncConsumer('localhost', 5672, 'ImageFinder', 'image_finder.size', 'myuser', 'mypassword')
     producer = RabbitMQProducer('localhost', 5672, 'myuser', 'mypassword')
